+++ src/refinement/stateful.rs
@@ -5,20 +5,11 @@ use serde::{Deserialize, Serialize};
 
 use crate::{Predicate, Refined, RefinementError, RefinementOps};
 
-use super::Refinement;
-
-/// An assertion that must hold for an instance of a type to be considered statefully refined.
-///
-/// Compared to [Predicate], the difference is that stateful predicates are "materialized" and
-/// may carry state along with them to be re-used across what would otherwise be independent
-/// tests.
 pub trait StatefulPredicate<T>: Default + Predicate<T> {
-    /// Whether a value satisfies the stateful predicate.
     fn test(&self, value: &T) -> bool {
         <Self as Predicate<T>>::test(value)
     }
 
-    /// An error message to display when the stateful predicate doesn't hold.
     fn error(&self) -> String {
         <Self as Predicate<T>>::error()
     }
@@ -28,7 +19,7 @@ pub trait StatefulPredicate<T>: Default + Predicate<T> {
 ///
 /// This is useful in situations where the additional space required to hold the
 /// predicate is outweighed by the cost of applying the predicate statelessly.
-/// Whenever possible, [Refinement](super::Refinement) should be preferred.
+/// Whenever possible, [Refinement] should be preferred.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
 #[cfg_attr(
     feature = "serde",
@@ -39,6 +30,26 @@ pub struct StatefulRefinement<T: Clone, P: StatefulPredicate<T> + Clone>(T, P);
 
 impl<T: Clone, P: StatefulPredicate<T> + Clone> RefinementOps for StatefulRefinement<T, P> {
     type T = T;
+
+    fn modify<F>(self, fun: F) -> Result<Self, RefinementError>
+    where
+        F: FnOnce(Self::T) -> Self::T,
+    {
+        let value = fun(self.0);
+        if self.1.test(&value) {
+            Ok(Self(value, self.1))
+        } else {
+            Err(RefinementError(self.1.error()))
+        }
+    }
+
+    fn replace(self, value: Self::T) -> Result<Self, RefinementError> {
+        if self.1.test(&value) {
+            Ok(Self(value, self.1))
+        } else {
+            Err(RefinementError(self.1.error()))
+        }
+    }
 
     fn extract(self) -> T {
         self.0
@@ -78,92 +89,5 @@ impl<T: Clone, P: StatefulPredicate<T> + Clone> TryFrom<Refined<T>> for Stateful
     }
 }
 
-impl<T: Clone, P: StatefulPredicate<T> + Clone> From<Refinement<T, P>>
-    for StatefulRefinement<T, P>
-{
-    fn from(value: Refinement<T, P>) -> Self {
-        Self(value.extract(), P::default())
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::*;
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn test_stateful_refinement_deserialize_success() {
-        let value =
-            serde_json::from_str::<StatefulRefinement<u8, boundable::unsigned::LessThan<5>>>("4")
-                .unwrap();
-        assert_eq!(*value, 4);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn test_stateful_refinement_deserialize_failure() {
-        let err =
-            serde_json::from_str::<StatefulRefinement<u8, boundable::unsigned::LessThan<5>>>("5")
-                .unwrap_err();
-        assert_eq!(
-            format!("{}", err),
-            "refinement violated: must be less than 5"
-        );
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn test_stateful_refinement_serialize() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(4, Default::default());
-        let serialized = serde_json::to_string(&value).unwrap();
-        assert_eq!(serialized, "4");
-    }
-
-    #[test]
-    fn test_stateful_refinement_modify_success() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(3, Default::default());
-        let modified = value.modify(|x| x + 1).unwrap();
-        assert_eq!(*modified, 4);
-    }
-
-    #[test]
-    fn test_stateful_refinement_modify_failure() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(4, Default::default());
-        let modified = value.modify(|x| x + 1).unwrap_err();
-        assert_eq!(
-            format!("{}", modified),
-            "refinement violated: must be less than 5"
-        );
-    }
-
-    #[test]
-    fn test_stateful_refinement_replace_success() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(4, Default::default());
-        let replaced = value.replace(3).unwrap();
-        assert_eq!(*replaced, 3);
-    }
-
-    #[test]
-    fn test_stateful_refinement_replace_failure() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(4, Default::default());
-        let replaced = value.replace(5).unwrap_err();
-        assert_eq!(
-            format!("{}", replaced),
-            "refinement violated: must be less than 5"
-        );
-    }
-
-    #[test]
-    fn test_stateful_refinement_extract() {
-        let value =
-            StatefulRefinement::<u8, boundable::unsigned::LessThan<5>>(4, Default::default());
-        let extracted = value.extract();
-        assert_eq!(extracted, 4);
-    }
-}
+mod tests {}